Changed the relationship between Kitchen and Cook + Added Swagger

+++ TheCodeKitchen.Server/TheCodeKitchen.Server.Core.Domain/Entities/MenuItem.cs
@@ -4,7 +4,7 @@ public class MenuItem
 {
     public double Price { get; set; }
     
-    public int MenuId { get; set; }
+    public long MenuId { get; set; }
     public Menu Menu { get; set; }
 
     public string ItemId { get; set; }
